horbert: init the hortbert workdir
We want to allow users to copy their files in the most simple way. For
this purpose we want to create an horber_workdir with one sub-directory
of each of the buttons. Each of the subdirectories will have a
number and a color (the same color as the Horbert buttons) so users will
know what to expect when their kid press each button.

Create the horbert_workdir at the path specified by the user. Create all
the sub0directories there based on a list.