horbert: convert and copy mp3 files
Internally, the Horbert is only able to read WAV files MP3 files
must be converted first. The WAV files must be copied on precise
directory names with precise file names:
- Directories on the SD card must be called 1, 2, 3, ..., 9. One
directory for each of the device buttons.
- Files within these directories must be called 1.WAV, 2.WAV, ...

For each on the horbert_workdir (where the user has put its MP3 files),
convert and name properly all those files within a .tmp directory.

Copy the content of each .tmp directory to the corresponding directory
on the SD card.

The conversion from mp3 to WAV is done with ffmpeg. It works properly
from the test we performed.